Refactor/cleanup
