+++ arpg-timeline/src/hoc/GameToSeasonWidget/types.ts
@@ -1,0 +1,9 @@
+"use client";
+
+import { Selector } from "./types";
+
+export type Selector = "current" | "next";
+export const SelectorLabels: Record<Selector, string> = {
+    current: "Current",
+    next: "The next",
+};
